shell: Extracdt WindowRenderer
