+++ shell/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod shell;
-pub use shell::{ApplicationContext, WindowRenderer};
+pub mod window_renderer;
+
+pub use shell::ApplicationContext;
+pub use window_renderer::WindowRenderer;
 
 pub fn time<T>(name: &str, f: impl FnOnce() -> T) -> T {
     let start = std::time::Instant::now();
